+++ src/sargeom/__init__.py
@@ -1,8 +1,10 @@
-from sargeom.trajectory import Trajectory, TRAJ_DTYPE
+from sargeom.trajectory import AntennaAttitude, NominalTrajectory, Trajectory, TRAJ_DTYPE
 
 __version__ = "0.4.0"
 
 __all__ = [
+    'AntennaAttitude',
+    'NominalTrajectory',
     'Trajectory',
     'TRAJ_DTYPE',
 ]